+++ spict/src/spict.cpp
@@ -158,6 +158,9 @@ Type objective_function<Type>::operator() ()
     DATA_SCALAR(simple);         // If simple=1 then use simple model (catch assumed known, no F process)
   DATA_SCALAR(dbg);            // Debug flag, if == 1 then print stuff.
   DATA_INTEGER(reportmode);    // If 1-5 only specific quantities are ADreported (increases speed, relevant for fitting within MSE)
+  DATA_INTEGER(simRandomEffects); // flag turning simulation of random effects on/off
+  DATA_INTEGER(simPriors);         // flag turning simulation of priors as random effects on/off
+
 
   // PARAMETERS
   PARAMETER_VECTOR(logm);      // m following the Fletcher formulation (see Prager 2002)
@@ -187,6 +190,9 @@ Type objective_function<Type>::operator() ()
   PARAMETER_VECTOR(SARvec);    // Autoregressive deviations to seasonal spline
   PARAMETER(logitSARphi);      // AR coefficient for seasonal spline dev
   PARAMETER(logSdSAR);         // Standard deviation seasonal spline deviations
+  // noise ratios as random effects for simulation with priors
+  PARAMETER_VECTOR(logalpha);  // log(sdi(i)/sdb)
+  PARAMETER(logbeta);          // log(sdc/sdf)
 
   //std::cout << "expmosc: " << expmosc(lambda, omega, 0.1) << std::endl;
   if(dbg > 0){
@@ -243,8 +249,7 @@ Type objective_function<Type>::operator() ()
   vector<Type> logq2(nq);
   for(int i=0; i<nq; i++){ logq2(i) = log(100) + logq(i); }
   //Type qf = exp(logqf);
-  Type n = exp(logn);
-  Type gamma = pow(n, n/(n-1.0)) / (n-1.0);
+
   Type lambda = exp(loglambda);
   Type delta = exp(logdelta);
   vector<Type> sdf(nsdf);
@@ -263,23 +268,9 @@ Type objective_function<Type>::operator() ()
   Type sdm = exp(logsdm);
   Type sdm2 = sdm*sdm;
   //Type isdm2 = 1.0/sdm2;
-  vector<Type> sdi = exp(logsdi);
-  vector<Type> sdi2(nsdi);
-  vector<Type> isdi2(nsdi);
-  for(int i=0; i<nsdi; i++){
-    sdi2(i) = sdi(i)*sdi(i);
-    isdi2(i) = 1.0/sdi2(i);
-  }
-  vector<Type> alpha = sdi/sdb;
-  vector<Type> logalpha = log(alpha);
   Type sde = exp(logsde);
   Type sde2 = sde*sde;
   Type isde2 = 1.0/sde2;
-  Type sdc = exp(logsdc);
-  Type sdc2 = sdc*sdc;
-  Type isdc2 = 1.0/sdc2;
-  Type beta = sdc/sdf(0);
-  Type logbeta = log(beta);
   Type SARphi = ilogit(logitSARphi);
   Type sdSAR = exp(logSdSAR);
 
@@ -308,6 +299,74 @@ Type objective_function<Type>::operator() ()
     //mvec(i) = exp(logm(0) + mu*logmcov(i) + logmre(i));
     mvec(i) = exp(logmc(i) + logmre(i));
   }
+
+  // Parameters with default priors are used in estimation of reference points
+  // Prior for logn
+  if(priorn(2) == 1){
+    ans -= dnorm(logn, priorn(0), priorn(1), 1);
+    if(simPriors == 1){
+      SIMULATE{
+        logn = rnorm(priorn(0), priorn(1));
+        REPORT(logn);
+      }
+    }
+  }
+  Type n = exp(logn);
+  Type gamma = pow(n, n/(n-1.0)) / (n-1.0);
+
+  vector<Type> sdi = exp(logsdi);
+  vector<Type> sdi2(nsdi);
+  vector<Type> isdi2(nsdi);
+  // Prior for logalpha
+  if(simPriors == 1){
+    if(prioralpha(2) == 1){
+      for(int i=0; i<nsdi; i++){
+        ans -= dnorm(logalpha(i), prioralpha(0), prioralpha(1), 1);
+        SIMULATE{
+          logalpha(i) = rnorm(prioralpha(0), prioralpha(1));
+        }
+        ans -= dnorm(logalpha(i), logsdi(i) - logsdb, Type(0.001), 1);
+      }
+      SIMULATE{
+        REPORT(logalpha);
+      }
+    }
+  }else{
+    vector<Type> alpha = sdi/sdb;
+    logalpha = log(alpha);
+    if(prioralpha(2) == 1){
+      for(int i=0; i<nsdi; i++){
+        ans -= dnorm(logalpha(i), prioralpha(0), prioralpha(1), 1);
+      }
+    }
+  }
+
+  for(int i=0; i<nsdi; i++){
+    sdi2(i) = sdi(i)*sdi(i);
+    isdi2(i) = 1.0/sdi2(i);
+  }
+
+  Type sdc = exp(logsdc);
+  Type sdc2 = sdc*sdc;
+  Type isdc2 = 1.0/sdc2;
+  // Prior for logbeta
+  if(simPriors == 1){
+    if(priorbeta(2) == 1){
+      ans -= dnorm(logbeta, priorbeta(0), priorbeta(1), 1);
+      SIMULATE{
+        logbeta = rnorm(priorbeta(0), priorbeta(1));
+        REPORT(logbeta);
+      }
+      ans -= dnorm(logbeta, logsdc - logsdf(0), Type(0.001), 1);
+    }
+  }else{
+    Type beta = sdc/sdf(0);
+    logbeta = log(beta);
+    if(priorbeta(2) == 1){
+      ans -= dnorm(logbeta, priorbeta(0), priorbeta(1), 1);
+    }
+  }
+
 
   Type p = n - 1.0;
   vector<Type> Bmsyd(nm);
@@ -514,9 +573,6 @@ Type objective_function<Type>::operator() ()
     ans-= dgamma(logn, priorngamma(0), 1.0/priorngamma(1), 1);
   }
   // Log-normal priors
-  if(priorn(2) == 1){
-    ans-= dnorm(logn, priorn(0), priorn(1), 1); // Prior for logn
-  }
   if((priorr(2) == 1) & (nm == 1)){
     ans-= dnorm(logr(0), priorr(0), priorr(1), 1); // Prior for logr
   }
@@ -561,14 +617,6 @@ Type objective_function<Type>::operator() ()
   }
   if(priorsdc(2) == 1){
     ans-= dnorm(logsdc, priorsdc(0), priorsdc(1), 1); // Prior for logsdc
-  }
-  if(prioralpha(2) == 1){
-    for(int i=0; i<nsdi; i++){
-      ans-= dnorm(logalpha(i), prioralpha(0), prioralpha(1), 1);  // Prior for logalpha
-    }
-  }
-  if(priorbeta(2) == 1){
-    ans-= dnorm(logbeta, priorbeta(0), priorbeta(1), 1); // Prior for logbeta
   }
   if(priorpsi(2) == 1){
     ans-= dnorm(logpsi, priorpsi(0), priorpsi(1), 1); // Prior for logsdm
@@ -622,10 +670,19 @@ Type objective_function<Type>::operator() ()
 
   using namespace density;
   ARk_t<Type> nldens(SARphivec);
-  if(seasontype==3) ans += SCALE(nldens, sdSAR)(vector<Type>(SARvec));
+  if(seasontype==3){
+    ans += SCALE(nldens, sdSAR)(vector<Type>(SARvec));
+    SIMULATE{
+      if(simRandomEffects == 1) nldens.simulate(SARvec);
+      REPORT(SARvec);
+    }
+  }
+  // std::cout << "-- sdf2: " << sdf2 << std::endl;
+  // std::cout << "-- sdf: " << sdf << std::endl;
 
   //vector<Type> logFs = logF
   vector<Type> logS(ns);
+  vector<Type> logFpred(ns);
   if(simple==0){
     if(dbg>0){
       std::cout << "--- DEBUG: F loop start --- ans: " << ans << std::endl;
@@ -641,13 +698,22 @@ Type objective_function<Type>::operator() ()
       if (efforttype == 2){
         Fpredtmp = predictF2(logF(i-1), dt(i), sdf2(iisdf), delta, logeta);
       }
-      Type logFpred = log( ffacvec(i) * Fpredtmp + fconvec(i) );
-      likval = dnorm(logF(i), logFpred, sqrt(dt(i-1))*sdf(iisdf), 1);
+      logFpred(i) = log( ffacvec(i) * Fpredtmp + fconvec(i) );
+      likval = dnorm(logF(i), logFpred(i), sqrt(dt(i-1)) * sdf(iisdf), 1);
       ans-=likval;
+      SIMULATE{
+        if(simRandomEffects == 1) logF(i) = rnorm(logFpred(i), sqrt(dt(i-1)) * sdf(iisdf));
+      }
+
       // DEBUGGING
       if(dbg>1){
         std::cout << "-- i: " << i << " -   logF(i-1): " << logF(i-1) << "  logF(i): " << logF(i) << "  ffacvec(i): " << ffacvec(i) << "  fconvec(i): " << fconvec(i) << "  iisdf: " << iisdf << "  sdf: " << sdf(iisdf) << "  efforttype: " << efforttype << "  likval: " << likval << "  ans:" << ans << std::endl;
       }
+    }
+    SIMULATE{
+      REPORT(logF);
+      vector<Type> trueF = exp(logFpred);
+      REPORT(trueF);
     }
 
     // Seasonal component
@@ -699,6 +765,9 @@ Type objective_function<Type>::operator() ()
           for(int k=0; k<logupred.size(); k++){
             if(dbg>0){ std::cout << "-- k: " << k << "- 2*j+k: " << 2*j+k << " - logu(2*j+k, i): " << logu(2*j+k, i) << std::endl; }
             likval += dnorm(logu(2*j+k, i), logupred(k), sduana, 1);
+            SIMULATE{
+              if(simRandomEffects == 1) logu(2*j+k, i) = rnorm(logupred(k), sduana);
+            }
           }
           ans-=likval;
           // DEBUGGING
@@ -710,11 +779,20 @@ Type objective_function<Type>::operator() ()
         for(int i=0; i<ns; i++) logS(i) += logu(2*j, i); // Sum diffusion and seasonal component
       }
     }
+    SIMULATE{
+      REPORT(logu);
+    }
+
   } else {
     for(int i=0; i<ns; i++) logS(i) = -30; // If using simple set fishing mortality to something small.
   }
   vector<Type> F = exp(logS + logF); // This is the fishing mortality used to calculate catch
   vector<Type> logFs = log(F);
+
+  SIMULATE{
+    REPORT(logS);
+    REPORT(logFs);
+  }
 
 
   // GROWTH RATE (modelled as time-varying m)
@@ -724,11 +802,17 @@ Type objective_function<Type>::operator() ()
     }
     // Compare initial value with stationary distribution of OU
     likval = dnorm(logmre(0), Type(0.0), sdm/sqrt(2.0*psi), 1);
+    SIMULATE{
+      if(simRandomEffects == 1) logmre(0) = rnorm(Type(0.0), sdm / sqrt(2.0 * psi));
+    }
     //likval = dnorm(logmre(0), logm(0), sdm/sqrt(2.0*psi), 1);
     ans -= likval;
     for (int i=1; i < ns; i++){
       Type logmrepred = predictm(logmre(i-1), dt(i-1), sdm2, psi);
       likval = dnorm(logmre(i), logmrepred, sqrt(dt(i-1))*sdm, 1);
+      SIMULATE{
+        if(simRandomEffects == 1) logmre(i) = rnorm(logmrepred, sqrt(dt(i-1)) * sdm);
+      }
       //likval = dnorm(logmre(i), logmre(i-1), sqrt(dt(i-1))*sdm, 1);
       ans -= likval;
       // DEBUGGING
@@ -736,7 +820,13 @@ Type objective_function<Type>::operator() ()
         std::cout << "-- i: " << i << " -   logmre(i-1): " << logmre(i-1) << "  sdm: " << sdm << "  likval: " << likval << "  ans:" << ans << std::endl;
       }
     }
+    SIMULATE{
+      REPORT(logmre);
+    }
   }
+
+  // std::cout << "-- sdb2: " << sdb2 << std::endl;
+  // std::cout << "-- sdb: " << sdb << std::endl;
 
   // BIOMASS PREDICTIONS
   if(dbg>0){
@@ -756,11 +846,22 @@ Type objective_function<Type>::operator() ()
       logFs(i) = logobsC(i) - logB(i); // Calculate fishing mortality
     }
     likval = dnorm(logBpred(i+1), logB(i+1), sqrt(dt(i))*sdb, 1);
+    SIMULATE{
+      if(simRandomEffects == 1){
+        logB(i+1) = rnorm(logBpred(i+1), sqrt(dt(i)) * sdb);
+        B(i+1) = exp(logB(i+1));
+      }
+    }
     ans-=likval;
     // DEBUGGING
     if(dbg>1){
       std::cout << "-- i: " << i << " -   logB(i+1): " << logB(i+1) << "  logBpred(i+1): " << logBpred(i+1) << "  sdb: " << sdb << "  likval: " << likval << "  ans:" << ans << std::endl;
     }
+  }
+  SIMULATE{
+    REPORT(logB);
+    vector<Type> trueB = exp(logBpred);
+    REPORT(trueB);
   }
   if(simple==1){ logFs(ns-1) = logFs(ns-2);}
 
@@ -830,14 +931,32 @@ Type objective_function<Type>::operator() ()
       inds = CppAD::Integer(isc(i)-1);
       if(robflagc==1){
         likval = log(pp*dnorm(logCpred(i), logobsC(i), stdevfacc(i)*sdc, 0) + (1.0-pp)*dnorm(logCpred(i), logobsC(i), robfac*stdevfacc(i)*sdc, 0));
+        SIMULATE{
+          Type uu = runif(0.0,1.0);
+          if(uu < pp){
+            logobsC(i) = log(rnorm(logCpred(i), stdevfacc(i) * sdc));
+          }else{
+            logobsC(i) = log(rnorm(logCpred(i), robfac * stdevfacc(i) * sdc));
+          }
+        }
+
       } else {
         likval = dnorm(logCpred(i), logobsC(i), stdevfacc(i)*sdc, 1);
+        SIMULATE{
+          logobsC(i) = rnorm(logCpred(i), stdevfacc(i) * sdc);
+        }
       }
       ans-= keep(inds) * likval;
       // DEBUGGING
       if(dbg>1){
         std::cout << "-- i: " << i << " -   logobsC(i): " << logobsC(i) << " -   stdevfacc(i): " << stdevfacc(i) << "  sdc: " << sdc << "  likval: " << likval << "  ans:" << ans << std::endl;
       }
+    }
+    SIMULATE{
+      vector<Type> obsC = exp(logobsC);
+      REPORT(obsC);
+      vector<Type> trueC = exp(logCpred);
+      REPORT(trueC);
     }
   }
 
@@ -850,8 +969,19 @@ Type objective_function<Type>::operator() ()
       logEpred(i) = logFcumpred(i) - logqf; // E = 1/q * integral{F_t dt}
       if(robflage==1){
         likval = log(pp*dnorm(logEpred(i), logobsE(i), stdevface(i)*sde, 0) + (1.0-pp)*dnorm(logEpred(i), logobsE(i), robfac*stdevface(i)*sde, 0));
+        SIMULATE{
+          Type uu = runif(0.0, 1.0);
+          if(uu < pp){
+            logobsE(i) = log(rnorm(logEpred(i), stdevface(i) * sde));
+          }else{
+            logobsE(i) = log(rnorm(logEpred(i), robfac * stdevface(i) * sde));  // same robfac and pp for C and E?
+          }
+        }
       } else {
         likval = dnorm(logEpred(i), logobsE(i), stdevface(i)*sde, 1);
+        SIMULATE{
+          logobsE(i) = rnorm(logEpred(i), stdevface(i) * sde);
+        }
       }
       inds = CppAD::Integer(ise(i)-1);
       ans-= keep(inds) * likval;
@@ -859,6 +989,12 @@ Type objective_function<Type>::operator() ()
       if(dbg>1){
         std::cout << "-- i: " << i << " -   logobsE(i): " << logobsE(i) << " -   stdevface(i): " << stdevface(i) << "  sde: " << sde << "  likval: " << likval << "  ans:" << ans << std::endl;
       }
+    }
+    SIMULATE{
+      vector<Type> obsE = exp(logobsE);
+      REPORT(obsE);
+      vector<Type> trueE = exp(logEpred);
+      REPORT(trueE);
     }
   }
 
@@ -877,14 +1013,48 @@ Type objective_function<Type>::operator() ()
     logIpred(i) = logq(indq) + log(B(ind));
     if(robflagi(indsdi)==1){
       likval = log(pp*dnorm(logobsI(i), logIpred(i), stdevfaci(i)*sdi(indsdi), 0) + (1.0-pp)*dnorm(logobsI(i), logIpred(i), robfac*stdevfaci(i)*sdi(indsdi), 0));
+      SIMULATE{
+        Type uu = runif(0.0,1.0);
+        if(uu < pp){
+          logobsI(i) = rnorm(logIpred(i), stdevfaci(i) * sdi(indsdi));
+        }else{
+          logobsI(i) = rnorm(logIpred(i), robfac * stdevfaci(i) * sdi(indsdi));
+        }
+      }
     } else {
       likval = dnorm(logobsI(i), logIpred(i), stdevfaci(i)*sdi(indsdi), 1);
+      SIMULATE{
+        logobsI(i) = rnorm(logIpred(i), stdevfaci(i) * sdi(indsdi));
+      }
     }
     ans-= keep(inds) * likval;
     // DEBUGGING
     if(dbg>1){
       std::cout << "-- i: " << i << " -  ind: " << ind << " -  indq: " << indq << " -  indsdi: " << indsdi << " -  inds: " << inds << " -   logobsI(i): " << logobsI(i) << "  logIpred(i): " << logIpred(i) << "  stdevfaci(i): " << stdevfaci(i) << "  likval: " << likval << "  sdi: " << sdi << "  ans:" << ans << std::endl;
     }
+  }
+  SIMULATE{
+    vector<Type> obsI = exp(logobsI);
+    REPORT(obsI);
+    vector<Type> trueI = exp(logIpred);
+    REPORT(trueI);
+  }
+
+  SIMULATE{
+    // Put obs back into obssrt
+    for(int i=0; i<nobsC; i++){
+      ind = CppAD::Integer(isc(i)-1);
+      obssrt(ind) = logobsC(i);
+    }
+    for(int i=0; i<nobsI; i++){
+      ind = CppAD::Integer(isi(i)-1);
+      obssrt(ind) = logobsI(i);
+    }
+    for(int i=0; i<nobsE; i++){
+      ind = CppAD::Integer(ise(i)-1);
+      obssrt(ind) = logobsE(i);
+    }
+    REPORT(obssrt);
   }
 
 
@@ -1140,11 +1310,11 @@ Type objective_function<Type>::operator() ()
   REPORT(MSY);
   REPORT(Bmsy);
   REPORT(Fmsy);
-  REPORT(stochmsy);
-  REPORT(logBBmsy);
-  REPORT(logFFmsy);
-  REPORT(logB);
-  REPORT(logF);
+  // REPORT(stochmsy);   // otherwise checkConsistency doesn't work
+  // REPORT(logBBmsy);
+  // REPORT(logFFmsy);
+  // REPORT(logB);
+  // REPORT(logF);
 
   return ans;
 }