+++ testmore/robust/script.R
@@ -90,6 +90,6 @@ plotspict.priors(fit2)
 plot(fit2)
 
 sink("res.out")
-print(round(sumspict.parest(fit),3))
-print(round(sumspict.parest(fit2),3))
+print(round(sumspict.parest(fit),2))
+print(round(sumspict.parest(fit2),2))
 sink()